# blog_system_backend/src/main/java/com/example/blog_system_backend/blog/BlogController.java
package com.example.blog_system_backend.blog;

import com.example.blog_system_backend.blog.dto.BlogCreateRequest;
import com.example.blog_system_backend.blog.dto.BlogResponse;
import com.example.blog_system_backend.blog.dto.BlogUpdateRequest;
import com.example.blog_system_backend.comment.CommentService;
import com.example.blog_system_backend.comment.dto.CommentRequest;
import com.example.blog_system_backend.comment.dto.CommentResponse;
import com.example.blog_system_backend.user.User;
import com.example.blog_system_backend.user.UserService;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;

import java.util.Collection;
import java.util.List;

@RestController
@RequestMapping("/api/blogs")
public class BlogController {

    private final BlogService blogService;
    private final UserService userService;
    private final CommentService commentService;

    public BlogController(BlogService blogService, UserService userService, CommentService commentService) {
        this.blogService = blogService;
        this.userService = userService;
        this.commentService = commentService;
    }

    @GetMapping("/{id}")
    public BlogResponse getById(@PathVariable Long id) {
        return blogService.getById(id);
    }

    @GetMapping
    public Page<BlogResponse> getAll(
            @RequestParam(required = false) Long categoryId,
            @RequestParam(required = false) Long tagId,
            @RequestParam(required = false) Collection<Long> tagIds,
            @RequestParam(required = false) String keyword,
            Pageable pageable
    ) {
        Collection<Long> resolvedTagIds = (tagIds != null && !tagIds.isEmpty())
                ? tagIds
                : (tagId != null ? List.of(tagId) : null);

        return blogService.findAll(categoryId, resolvedTagIds, keyword, pageable);
    }

    @PostMapping
    public BlogResponse create(
            @RequestBody BlogCreateRequest request,
            Authentication authentication
    ) {
        String username = authentication.getName();

        User currentUser = userService.findUserByUsername(username);
        System.out.println("=====================> 拿到的 userId: " + currentUser.getId());
        return blogService.create(request, currentUser.getId());
    }

    @PutMapping("/{id}")
    public BlogResponse update(
            @PathVariable Long id,
            @RequestBody BlogUpdateRequest request,
            Authentication authentication
    ) {
        String username = authentication.getName();

        User currentUser = userService.findUserByUsername(username);
        return blogService.update(id, request, currentUser.getId());
    }

    @DeleteMapping("/{id}")
    public void delete(
            @PathVariable Long id,
            Authentication authentication
    ) {
        String username = authentication.getName();

        User currentUser = userService.findUserByUsername(username);
        blogService.delete(id, currentUser.getId());
    }

    // 获取某博客下的所有评论
    @GetMapping("/{blogId}/comments")
    public Page<CommentResponse> getCommentsByBlogId(
            @PathVariable Long blogId,
            Pageable pageable
    ) {
        return commentService.getByBlogId(blogId, pageable);
    }

    // 发表顶级评论
    @PostMapping("/{blogId}/comments")
    public CommentResponse create(
            @PathVariable Long blogId,
            @RequestBody CommentRequest request,
            Authentication authentication
    ) {
        String username = authentication.getName();
        User currentUser = userService.findUserByUsername(username);

        System.out.println("=====================> 评论用户 ID: " + currentUser.getId());
        return commentService.create(blogId, currentUser.getId(), request);
    }

    // 点赞 / 取消点赞
    @PostMapping("/{blogId}/like")
    public boolean toggleLike(
            @PathVariable Long blogId,
            Authentication authentication
    ) {
        String username = authentication.getName();
        User currentUser = userService.findUserByUsername(username);
        Long userId = currentUser.getId();

        return blogService.toggleLike(blogId, userId);
    }

    // 记录浏览量
    @PostMapping("/{id}/view")
    public int recordView(@PathVariable Long id) {
        return blogService.incrementViewCount(id);
    }

    // 查询当前用户是否点赞
    @GetMapping("/{blogId}/isLiked")
    public boolean isLiked(
            @PathVariable Long blogId,
            Authentication authentication
    ) {
        String username = authentication.getName();
        User currentUser = userService.findUserByUsername(username);
        Long userId = currentUser.getId();

        return blogService.isLiked(blogId, userId);
    }
}